Use new errorMessage props of text-input component.
- oc-text-input now has an errorMessage property for showing an error
message in a fixed way.
- also use prop of oc-text-input to always reserve vertical space for
the error message.
- Debounce error message changes:
The dialog reacted to error messages as soon as they were created which
can be annoying.
For example there is an error message when the name of a new folder ends
in a blank. If you type folder names containing blanks, the error
message would appear during typing for every single blank for a brief
moment and then would disappear when the next non-blank char is typed.
Debouncing error messages causes that they don't appear during typing
but only if you stop typing for the specified time (400 milliseconds).